refactor(hooks): name the counter for what it counts
bg-agent-counter counted no agent call at all - no agent payload carries
run_in_background, which is the defect GH-92 was filed on - and every
reader of the three hooks met the wrong name. It is now
background-call-counter, with its state file renamed to match, and the
old installed path is recorded as retired so a machine whose manifest was
lost is told about the leftover rather than left with it live.
