+++ README.md
@@ -26,7 +26,7 @@ pipeline of persona agents and commands built on top of them.
 - `skill-gate.js` — on the same write payloads, and on any command that writes a file (redirection, `tee`, `Set-Content`/`Add-Content`/`Out-File`, `cp`/`mv` destination), names the skills that claim the written file (their `paths:` frontmatter) plus the ones they list in `with:`, minus whatever the agent making the call has already loaded; the first offending call is denied until the skills are loaded, a repeat of the same file and missing set only warns. A subagent is gated on its own loads, read from its own transcript, and the main agent's loads do not vouch for it
 
 **`PostToolUse`** — runs after a tool call, on two routes read off the payload the same way: a path arriving with the content written to it, or one of `Edit` / `Write` / `MultiEdit` / `NotebookEdit` naming a path with no content field, reaches the write tools; a call marked `run_in_background: true` reaches the counter. A path named for reading reaches nothing. The three lint tools run only on C++ files (`.h`, `.hpp`, `.cpp`, `.cc`, `.cxx`):
-- `bg-agent-counter.js` — counts a call started in the background, whichever tool made it, as work the `Stop` notification below must wait for. Such a call returns the moment it is launched, so this runs while the work is still going; a call refused at any gate never reaches this hook, and is never counted
+- `background-call-counter.js` — counts a call started in the background, whichever tool made it, as work the `Stop` notification below must wait for. Such a call returns the moment it is launched, so this runs while the work is still going; a call refused at any gate never reaches this hook, and is never counted
 - `comment-check.js` — runs on every file whose comment syntax it knows (`//`, `/* */`, `#`, `--`, `;`, `%`, `<!-- -->`), no config file needed; reminds to check any newly added non-doc comment against the `comments` skill
 - `clang-format.js` — formats in-place; requires `.clang-format` in this repository (searched from the edited file up to the git root, never into a parent repo) — skips silently if missing, or if `clang-format` isn't in PATH
 - `clang-tidy.js` — runs static analysis; requires `.clang-tidy` (uses `compile_commands.json` when found)
